+++ GRDB/Core/DatabasePool.swift
@@ -312,6 +312,32 @@ extension DatabasePool {
 
 
 // =========================================================================
+// MARK: - Transaction Observers
+
+extension DatabasePool {
+    
+    /// Add a transaction observer, so that it gets notified of all
+    /// database changes.
+    ///
+    /// Database holds weak references to its transaction observers: they are
+    /// not retained, and stop getting notifications after they are deallocated.
+    public func addTransactionObserver(transactionObserver: TransactionObserverType) {
+        writer.inDatabase { db in
+            db.addTransactionObserver(transactionObserver)
+        }
+    }
+    
+    /// Remove a transaction observer.
+    public func removeTransactionObserver(transactionObserver: TransactionObserverType) {
+        writer.inDatabase { db in
+            db.removeTransactionObserver(transactionObserver)
+        }
+    }
+
+}
+
+
+// =========================================================================
 // MARK: - DatabaseReader
 
 extension DatabasePool : DatabaseReader {